[#1770] Added capital spend import
