+++ akvo/iati/imports/fields/__init__.py
@@ -4,7 +4,6 @@
 # See more details in the license.txt file located at the root folder of the Akvo RSR module.
 # For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
 
-from .budget_items import budget_items, country_budget_items
 from .conditions import conditions
 from .contacts import contacts
 from .classifications import policy_markers, sectors
@@ -13,23 +12,24 @@ from .defaults import (currency, default_aid_type, default_finance_type, default
                        default_tied_status, hierarchy, language, scope)
 from .descriptions import (background, current_status, goals_overview, project_plan,
                            project_plan_summary, sustainability, target_group)
+from .financials import (budget_items, capital_spend, country_budget_items, planned_disbursements,
+                         transactions)
 from .legacy_data import legacy_data
 from .links import current_image, documents, links
 from .locations import locations, recipient_countries, recipient_regions
 from .partnerships import partnerships
-from .planned_disbursements import planned_disbursements
 from .related_projects import related_projects
 from .results import results
 from .special_reporting import crs_add, fss
 from .status import status
 from .titles import title, subtitle
-from .transactions import transactions
 
 __all__ = [
     'actual_end_date',
     'actual_start_date',
     'background',
     'budget_items',
+    'capital_spend',
     'conditions',
     'contacts',
     'country_budget_items',
